requirements.txt
openai
python-dotenv
streamlit==1.29
PyPDF2
langchain